Follow coding standard
