+++ src/main/java/duke/task/Event.java
@@ -1,7 +1,5 @@
 package duke.task;
 
-import duke.task.Task;
-
 public class Event extends Task {
 
     protected String at;
